refactor the types from dbs package to event package
