+++ code/go/0chain.net/smartcontract/dbs/event/types.go
@@ -1,4 +1,4 @@
-package dbs
+package event
 
 type DbUpdates struct {
 	Id      string                 `json:"id"`
