+++ code/go/0chain.net/smartcontract/storagesc/stakepool.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 
-	"0chain.net/smartcontract/dbs"
 	"0chain.net/smartcontract/stakepool/spenum"
 
 	"0chain.net/smartcontract/dbs/event"
@@ -461,7 +460,7 @@ func (ssc *StorageSmartContract) stakePoolLock(t *transaction.Transaction,
 	}
 
 	// TO-DO: Update stake in eventDB
-	data, _ := json.Marshal(dbs.DbUpdates{
+	data, _ := json.Marshal(event.DbUpdates{
 		Id: spr.BlobberID,
 		Updates: map[string]interface{}{
 			"total_stake": int64(sp.stake()),
@@ -516,7 +515,7 @@ func (ssc *StorageSmartContract) stakePoolUnlock(
 			"saving stake pool: %v", err)
 	}
 
-	data, _ := json.Marshal(dbs.DbUpdates{
+	data, _ := json.Marshal(event.DbUpdates{
 		Id: spr.BlobberID,
 		Updates: map[string]interface{}{
 			"total_stake": int64(sp.stake()),
